+++ src/main/main.cpp
@@ -38,10 +38,17 @@ main(int32_t argc, char* argv[])
     {
       options.set(bitwuzla::Option::PRODUCE_UNSAT_CORES, 1);
     }
+    if (main_options.print_model)
+    {
+      options.set(bitwuzla::Option::PRODUCE_MODELS, 1);
+    }
 
     std::cout << bitwuzla::set_bv_format(main_options.bv_format);
-    bitwuzla::parser::Parser parser(
-        tm, options, main_options.language, &std::cout);
+    bitwuzla::parser::Parser parser(tm,
+                                    options,
+                                    main_options.language,
+                                    &std::cout,
+                                    main_options.print_model);
     parser.parse(main_options.infile_name,
                  main_options.print || main_options.parse_only);
     reset_time_limit();