+++ src/main/options.h
@@ -21,6 +21,7 @@ struct Options
 {
   bool print              = false;
   bool print_unsat_core   = false;
+  bool print_model        = false;
   bool parse_only         = false;
   uint8_t bv_format       = 2;
   uint64_t time_limit     = 0;